prediction_api/requirements.txt
Flask
Pillow
numpy
tflite-runtime  # Plus léger que tensorflow
# Ou si vous préférez garder le package complet :
# tensorflow